Fixes in GUI for fucked up PyQt

+++ gui/utils/matplotlib.py
@@ -86,7 +86,7 @@ class FigureCanvasWithInfo(FigureCanvas):
         if rows == 0:
             self.info_table.setVisible(False)
         else:
-            self.info_table.setFixedHeight(self.info_table.sizeHintForRow(0) * rows + 4)
+            self.info_table.setFixedHeight(int(self.info_table.sizeHintForRow(0) * rows + 4))
             self.info_table.setVisible(True)
 
     def _on_select_info(self, current, _):
